% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pairwise_mmds.R
\name{calc_pairwise_mmds}
\alias{calc_pairwise_mmds}
\title{calc_pairwise_mmds}
\usage{
calc_pairwise_mmds(
  qc_dt,
  qc_names = c("log_counts", "log_feats", "logit_mito"),
  annots_disc = NULL,
  annots_cont = NULL,
  n_cores = 4,
  sigma = length(qc_names),
  subsample = 100,
  n_times = 20,
  seed = 22
)
}
\arguments{
\item{qc_dt}{data.table of QC metrics for all cells and samples}

\item{qc_names}{List of metrics to actually use for calculating
sample-to-sample distances}

\item{annots_disc, annots_cont}{Which annotating variables should be added
to the SampleQC object? These should be entries in qc_dt. annots_disc
indicates discrete variables; annots_cont continuous variables.}

\item{n_cores}{How many cores to parallelize over?}

\item{sigma}{Scale for MMD kernel (defaults to length of qc_names)}

\item{subsample}{Should we downsample the number of cells per sample to
this number}

\item{n_times}{How many times do we sample MMD between each pair of samples?
(if subsampled, MMD is a stochastic value)}
}
\value{
slightly funky \code{sce} object
}
\description{
Maximum mean discrepancy (MMD) is a method for measuring distances between
multivariate distributions. \emph{SampleQC} uses this to determine which
samples have similar distributions, and should therefore have the same
model fit to them. In addition
}
\details{
Calculates sample-to-sample MMD distances, clusters and embeds them

The most important variable to consider here is \emph{qc_names}, which
specifies the set of QC metrics that will be used both here and in later
steps. The default is to use log counts, log features and the logit
transformation of the mitochondrial proportion (under this transformation,
the distribution is reasonably modelled as a gaussian mixture model)
}
