% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.add_annot_vars}
\alias{.add_annot_vars}
\title{Adds user-specified annotation variables}
\usage{
.add_annot_vars(qc_dt, qc_df, annot_vars)
}
\arguments{
\item{qc_dt}{data.table}

\item{qc_df}{data.frame object containing calculated QC metrics}
}
\value{
qc_dt, a data.table containing the sample variable plus qc metrics
}
\description{
Adds user-specified annotation variables
}
\keyword{internal}
