% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{.add_annotations}
\alias{.add_annotations}
\title{Checks that input is ok, puts it into expected format}
\usage{
.add_annotations(qc_dt)
}
\arguments{
\item{sce}{SingleCellExperiment or data.frame (or data.table, some other
class inheriting data.frame) containing calculated QC metrics}

\item{qc_names}{list of qc_names that need to be extracted}
}
\value{
qc_dt, a data.table containing the sample variable plus qc metrics
}
\description{
Checks that input is ok, puts it into expected format
}
\keyword{internal}
