% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.add_log_feats}
\alias{.add_log_feats}
\title{Add log feats to qc_dt}
\usage{
.add_log_feats(qc_dt, qc_df)
}
\arguments{
\item{qc_dt}{data.table of QC metrics}

\item{qc_df}{input data}
}
\description{
Add log feats to qc_dt
}
\keyword{internal}
