% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.add_logit_mito}
\alias{.add_logit_mito}
\title{Add logit-transformed mitochondrial proportions to qc_dt}
\usage{
.add_logit_mito(qc_dt, qc_df)
}
\arguments{
\item{qc_dt}{data.table of QC metrics}

\item{qc_df}{input data}
}
\description{
Add logit-transformed mitochondrial proportions to qc_dt
}
\keyword{internal}
