% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.add_qc_annots}
\alias{.add_qc_annots}
\title{Adds sample-level annotations for each known QC metric}
\usage{
.add_qc_annots(qc_dt)
}
\arguments{
\item{qc_dt}{data.table}
}
\value{
qc_dt, a data.table containing the sample variable plus qc metrics
}
\description{
Adds sample-level annotations for each known QC metric
}
\keyword{internal}
