% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.add_splice_ratio}
\alias{.add_splice_ratio}
\title{Add log splice ratio to qc_dt}
\usage{
.add_splice_ratio(qc_dt, qc_df)
}
\arguments{
\item{qc_dt}{data.table of QC metrics}

\item{qc_df}{input data}
}
\description{
Add log splice ratio to qc_dt
}
\keyword{internal}
