% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.add_unknown_metrics}
\alias{.add_unknown_metrics}
\title{Adds metrics that SampleQC doesn't have specific functions for}
\usage{
.add_unknown_metrics(qc_dt, qc_df, qc_names)
}
\arguments{
\item{qc_dt}{data.table of QC metrics}

\item{qc_df}{data.frame object containing calculated QC metrics}

\item{qc_names}{list of qc_names that need to be extracted}
}
\description{
Adds metrics that SampleQC doesn't have specific functions for
}
\keyword{internal}
