% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.assemble_fit_results}
\alias{.assemble_fit_results}
\title{Puts results from fitting models into SingleCellExperiment format}
\usage{
.assemble_fit_results(qc_obj, fit_list, fit_params)
}
\value{
qc_obj with things added
}
\description{
Puts results from fitting models into SingleCellExperiment format
}
\keyword{internal}
