% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.calc_ellipses_dt}
\alias{.calc_ellipses_dt}
\title{Calculates path for an ellipse based on 2D mu and sigma values}
\usage{
.calc_ellipses_dt(qc_obj, sel_sample, dim1, dim2, dim2_name, alpha = 0.01)
}
\arguments{
\item{qc_obj}{Output from function fit_sampleQC}

\item{sel_sample}{Which sample?}

\item{dim1, dim2}{Dimensions to use}

\item{dim2_name}{Name of second dimension}

\item{alpha}{Chi-squared quantile to use for ellipse sizes}
}
\value{
data.table object
}
\description{
Calculates path for an ellipse based on 2D mu and sigma values
}
\keyword{internal}
