% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{.calc_mmd_mat}
\alias{.calc_mmd_mat}
\title{Calculates mean MMD distances between all pairs of samples}
\usage{
.calc_mmd_mat(sample_list, mat_list, n_times, subsample, sigma, n_cores)
}
\arguments{
\item{sample_list}{List of samples to iterate through}

\item{mat_list}{List of matrices of QC metrics}

\item{n_times}{How many times do we sample MMD between each pair of samples?
(if subsampled, MMD is a stochastic value)}

\item{subsample}{Should we downsample the number of cells per sample to
this number}

\item{sigma}{Scale for MMD kernel (defaults to length of qc_names)}

\item{n_cores}{How many cores to parallelize over}
}
\value{
matrix of MMD values
}
\description{
Calculates mean MMD distances between all pairs of samples
}
\keyword{internal}
