% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.calc_outliers_dt}
\alias{.calc_outliers_dt}
\title{Calculate outliers for each cell}
\usage{
.calc_outliers_dt(em_list, x, groups, alpha)
}
\arguments{
\item{em_list}{Most of output from .fit_one_sampleQC}

\item{x}{Matrix of QC values}

\item{groups}{Sample labels for every cell}

\item{alpha}{Chi squared threshold defining outliers}
}
\value{
data.table containing fitted parameters
}
\description{
Calculate outliers for each cell
}
\keyword{internal}
