% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.calc_outliers_plot}
\alias{.calc_outliers_plot}
\title{Make data.table for plotting outliers over QC metrics}
\usage{
.calc_outliers_plot(qc_dt, em_list)
}
\arguments{
\item{qc_dt}{data.table of QC metrics}

\item{em_list}{Most of output from .fit_one_sampleQC}
}
\value{
data.table containing fitted parameters
}
\description{
Make data.table for plotting outliers over QC metrics
}
\keyword{internal}
