% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.calc_p_jk_proportions}
\alias{.calc_p_jk_proportions}
\title{Calculate proportions of cells within each sample allocated to each}
\usage{
.calc_p_jk_proportions(em_list, x, groups, alpha)
}
\arguments{
\item{em_list}{Most of output from .fit_one_sampleQC}

\item{x}{Matrix of QC values}

\item{groups}{Sample labels for every cell}

\item{alpha}{Chi squared threshold defining outliers}
}
\value{
data.table containing fitted parameters
}
\description{
Calculate proportions of cells within each sample allocated to each
}
\keyword{internal}
