% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.check_fit_params}
\alias{.check_fit_params}
\title{Checks that the parameters specified are ok}
\usage{
.check_fit_params(
  qc_obj,
  K_all,
  K_list,
  n_cores,
  alpha,
  em_iters,
  mcd_alpha,
  mcd_iters,
  method,
  track
)
}
\description{
Checks that the parameters specified are ok
}
\keyword{internal}
