% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{.check_is_qc_obj}
\alias{.check_is_qc_obj}
\title{Checks whether a given sce object is a SampleQC object}
\usage{
.check_is_qc_obj(qc_obj)
}
\arguments{
\item{qc_obj}{SampleQC object}
}
\value{
string saying progress
}
\description{
Checks whether a given sce object is a SampleQC object
}
