% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{.check_qc_dt}
\alias{.check_qc_dt}
\title{Checks that input is ok, puts it into expected format}
\usage{
.check_qc_dt(qc_dt, qc_names)
}
\arguments{
\item{qc_dt}{data.table}
}
\description{
Checks that input is ok, puts it into expected format
}
\keyword{internal}
