% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pairwise_mmds.R
\name{.cluster_mmd_graph}
\alias{.cluster_mmd_graph}
\title{Cluster igraph object using Louvain clustering}
\usage{
.cluster_mmd_graph(mmd_graph, one_group_only)
}
\arguments{
\item{mmd_graph}{}
}
\value{
vector of group_ids for all samples
}
\description{
Cluster igraph object using Louvain clustering
}
\keyword{internal}
