% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{.dist_fn}
\alias{.dist_fn}
\title{Wrapper for cpp MMD distance function}
\usage{
.dist_fn(m_i, m_j, subsample, sigma)
}
\arguments{
\item{m_i}{matrix i}

\item{m_j}{matrix j}

\item{subsample}{Number to subsample down to}

\item{sigma}{Number to subsample down to}
}
\value{
MMD between subsamples of input matrices
}
\description{
Wrapper for cpp MMD distance function
}
\keyword{internal}
