% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.draw_alpha_j}
\alias{.draw_alpha_j}
\title{Draws random parameters for sample shifts}
\usage{
.draw_alpha_j(D, J)
}
\arguments{
\item{D}{number of dimensions}

\item{J}{number of samples}
}
\value{
matrix of J rows by D columns
}
\description{
Draws random parameters for sample shifts
}
\keyword{internal}
