% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.draw_delta_jk}
\alias{.draw_delta_jk}
\title{Draws random parameters for mixture model components}
\usage{
.draw_delta_jk(D, J, K)
}
\arguments{
\item{D}{number of components}

\item{J}{number of groups}

\item{K}{number of components}
}
\value{
?
}
\description{
Draws random parameters for mixture model components
}
\keyword{internal}
