% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_expt_params}
\alias{.draw_expt_params}
\title{Draws experiment-level parameters}
\usage{
.draw_expt_params(n_groups, n_cells, cells_p_s, D, K, qc_names, sel_ks = NULL)
}
\arguments{
\item{n_groups}{Number of sample groups, each of which is composed of
multiple samples}

\item{n_cells}{Total number of cells in experiment}

\item{cells_p_s}{Average number of cells per sample (not per
sample group)}

\item{D}{Number of dimensions}

\item{K}{Number of components}

\item{qc_names}{Names for qc metrics}
}
\value{
list of parameters
}
\description{
Draws experiment-level parameters
}
\keyword{internal}
