% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_mu_0s}
\alias{.draw_mu_0s}
\title{Draws centre of sample group}
\usage{
.draw_mu_0s(D, n_groups)
}
\arguments{
\item{D}{number of dimensions}

\item{n_groups}{number of sample groups}
}
\value{
vector of D columns
}
\description{
Draws centre of sample group
}
\keyword{internal}
