% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.draw_out_j}
\alias{.draw_out_j}
\title{Draws random parameters to determine how to do outliers}
\usage{
.draw_out_j(J, p_out_0, theta_0, p_loss_0)
}
\arguments{
\item{J}{number of samples}
}
\value{
\code{matrix} with rows corresponding to samples
col 1 is \code{p_out}, proportion of each sample which is an outlier
col 2 is \code{p_lost}, mean proportion of non-mito counts lost in outliers
}
\description{
Draws random parameters to determine how to do outliers
}
\keyword{internal}
