% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.draw_outliers}
\alias{.draw_outliers}
\title{Which cells are outliers?}
\usage{
.draw_outliers(samples, out_j, N)
}
\arguments{
\item{samples}{sample indices}

\item{out_j}{probabilities of being outliers}

\item{N}{number of cells}
}
\value{
vector of outlier status
}
\description{
Which cells are outliers?
}
\keyword{internal}
