% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.draw_p_jk}
\alias{.draw_p_jk}
\title{Draws random parameters for mixing parameters}
\usage{
.draw_p_jk(J, K, dir_0)
}
\arguments{
\item{J}{number of samples}

\item{K}{number of components}

\item{dir_0}{parameters for Dirichlet distribution}
}
\value{
?
}
\description{
Draws random parameters for mixing parameters
}
\keyword{internal}
