% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_p_loss_0s}
\alias{.draw_p_loss_0s}
\title{Draws group-level probabilities of reads being lost in outlier cells}
\usage{
.draw_p_loss_0s(n_groups)
}
\arguments{
\item{n_groups}{number of sample groups}
}
\value{
vector of probabilities, one for each sample group
}
\description{
Draws group-level probabilities of reads being lost in outlier cells
}
\keyword{internal}
