% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_samples}
\alias{.draw_samples}
\title{Randomly splits up N cells into J samples with different sizes}
\usage{
.draw_samples(J, N)
}
\arguments{
\item{J}{number of samples}

\item{N}{number of cells}
}
\value{
vector of samples
}
\description{
Randomly splits up N cells into J samples with different sizes
}
\keyword{internal}
