% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.draw_sel_ks}
\alias{.draw_sel_ks}
\title{Determines which components are observed in each sample group}
\usage{
.draw_sel_ks(K, n_groups)
}
\arguments{
\item{K}{number of components}

\item{n_groups}{number of sample groups}
}
\value{
matrix of 0s and 1s, where rows correspond to sample groups, and
columns correspond to components.
}
\description{
Determines which components are observed in each sample group
}
\keyword{internal}
