% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_Sigma_k}
\alias{.draw_Sigma_k}
\title{Draws random parameters for mixture model covariances}
\usage{
.draw_Sigma_k(D, K)
}
\arguments{
\item{D}{number of dimensions}

\item{K}{number of components}
}
\value{
?
}
\description{
Draws random parameters for mixture model covariances
}
\keyword{internal}
