% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_theta_0s}
\alias{.draw_theta_0s}
\title{Sample group-level probabilities of cells being outliers}
\usage{
.draw_theta_0s(n_groups)
}
\arguments{
\item{n_groups}{number of sample groups}
}
\value{
vector of theta values for beta-binomial distributions, one for each sample group
}
\description{
Sample group-level probabilities of cells being outliers
}
\keyword{internal}
