% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_z}
\alias{.draw_z}
\title{Draws latent true component for each cell}
\usage{
.draw_z(samples, p_jk, N, J, K)
}
\arguments{
\item{samples}{sample indices}

\item{p_jk}{component probabilities}

\item{N}{number of cells}

\item{J}{number of samples}

\item{K}{number of components}
}
\value{
?
}
\description{
Draws latent true component for each cell
}
\keyword{internal}
