% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{.embed_mds}
\alias{.embed_mds}
\title{MDS embedding of MMD distances}
\usage{
.embed_mds(mmd_mat)
}
\arguments{
\item{mmd_mat}{matrix of MMD distances}
}
\value{
matrix of embedding
}
\description{
MDS embedding of MMD distances
}
