% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pairwise_mmds.R
\name{.embed_umap}
\alias{.embed_umap}
\title{Calculates non-linear embeddings of MMD distances calculated by
calc_pairwise_mmds}
\usage{
.embed_umap(mmd_mat, n_nhbrs)
}
\arguments{
\item{mmd_mat}{MMD distances}

\item{n_nhbrs}{How many neighbours for UMAP nearest neighbours graph}
}
\value{
matrix of embedding
}
\description{
Calculates non-linear embeddings of MMD distances calculated by
calc_pairwise_mmds
}
\keyword{internal}
