% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sampleqc.R
\name{.fit_one_sampleqc}
\alias{.fit_one_sampleqc}
\title{Fits \code{SampleQC} model to one cluster of samples}
\usage{
.fit_one_sampleqc(df, K = 1, fit_params, seed)
}
\arguments{
\item{df}{DataFrame with lots of nested things}

\item{K}{How many QC celltypes are there?}

\item{fit_params}{List of parameters specifying the model}
}
\value{
list, containing lots of cell outlier information
}
\description{
Fits \code{SampleQC} model to one cluster of samples
}
\keyword{internal}
