% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{.init_qc_dt}
\alias{.init_qc_dt}
\title{Initializes qc_dt object}
\usage{
.init_qc_dt(qc_df, sample_var)
}
\arguments{
\item{qc_df}{input data}

\item{sample_var}{which column of df has sample labels? (e.g. sample, group,
batch, library)}
}
\description{
Initializes qc_dt object
}
\keyword{internal}
