% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{.make_mmd_graph}
\alias{.make_mmd_graph}
\title{Build igraph object from MMD matrix}
\usage{
.make_mmd_graph(mmd_mat, n_nhbrs = 5)
}
\arguments{
\item{mmd_mat}{matrix of MMD distances}

\item{n_nhbrs}{How many neighbours in the NN graph}
}
\value{
igraph object
}
\description{
Build igraph object from MMD matrix
}
\keyword{internal}
