% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.make_mu_jk_dt}
\alias{.make_mu_jk_dt}
\title{Extract parameters from fitted em_list}
\usage{
.make_mu_jk_dt(em_list)
}
\arguments{
\item{em_list}{Most of output from .fit_one_sampleQC}
}
\value{
data.table containing fitted parameters
}
\description{
Extract parameters from fitted em_list
}
\keyword{internal}
