% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.make_toy_qc_df}
\alias{.make_toy_qc_df}
\title{Generate fake qc_df object for testing}
\usage{
.make_toy_qc_df()
}
\value{
data.frame
}
\description{
Generate fake qc_df object for testing
}
\keyword{internal}
