% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{.plot_one_embedding}
\alias{.plot_one_embedding}
\title{Plots non-linear embeddings of MMD distances calculated by
embed_sample_to_sample_MMDs}
\usage{
.plot_one_embedding(mmd_list, v, var_type, sel_embed)
}
\arguments{
\item{var_type}{Is this annotation discrete or continuous?}

\item{sel_embed}{Which type of embedding to use}

\item{embed_dt}{Output from embed_sample_to_sample_MMDs}

\item{annot_var}{Which annotation variable to plot}
}
\value{
ggplot object
}
\description{
Plots non-linear embeddings of MMD distances calculated by
embed_sample_to_sample_MMDs
}
\keyword{internal}
