% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.process_outputs}
\alias{.process_outputs}
\title{Gathers results simulated for individual sample groups together}
\usage{
.process_outputs(group_sims, expt_params)
}
\arguments{
\item{group_sims}{List of results for individual sample groups}

\item{expt_params}{True parameters used to generate group_sims}
}
\value{
\code{list} of outputs
}
\description{
Gathers results simulated for individual sample groups together into
results for a whole experiment. Results for individual sample groups are
returned.
}
\keyword{internal}
