% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.sim_outliers}
\alias{.sim_outliers}
\title{Simulates outliers from ok cells
within each sample, downsample non-mito
then recalculate metrics

not implemented: decide whether to downsample non-mito or \emph{all};}
\usage{
.sim_outliers(x_ok, samples, outliers, out_j, J)
}
\arguments{
\item{x_ok}{healthy cells}

\item{samples}{list of sample membership}

\item{out_j}{parameters for sample outliers}

\item{J}{number of samples}
}
\value{
\code{matrix} of cell QC metrics including outliers
}
\description{
Simulates outliers from ok cells
within each sample, downsample non-mito
then recalculate metrics

not implemented: decide whether to downsample non-mito or \emph{all};
}
\keyword{internal}
