% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.sim_sample_group}
\alias{.sim_sample_group}
\title{Simulates QC metrics for one group of samples}
\usage{
.sim_sample_group(
  N,
  D,
  J,
  K,
  mu_0,
  beta_k,
  Sigma_k,
  p_out_0,
  theta_0,
  p_loss_0
)
}
\arguments{
\item{D}{number of dimensions}

\item{J}{number of groups}

\item{K}{number of components}
}
\value{
list with two entries:
\itemize{
\item \code{qc_dt}, a \code{\link[data.table]{data.table}}
\item \code{params}, a list specifying the true parameter values for this group
}
}
\description{
Simulates QC metrics for one group of samples
}
\section{Details}{

}

\keyword{internal}
