% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{embed_sample_to_sample_MMDs}
\alias{embed_sample_to_sample_MMDs}
\title{Calculates non-linear embeddings of MMD distances calculated by
calculate_sample_to_sample_MMDs}
\usage{
embed_sample_to_sample_MMDs(
  mmd_list,
  qc_dt,
  annot_discrete = NULL,
  annot_cont = NULL,
  n_nhbrs = 5
)
}
\arguments{
\item{mmd_list}{Outputs from function calculate_sample_to_sample_MMDs}

\item{qc_dt}{Data.table of QC metrics for all cells and samples}

\item{annot_discrete}{List of discrete-valued variables to include for plotting}

\item{annot_cont}{List of continuous-valued variables to include for plotting}

\item{n_nhbrs}{How many neighbours in nearest neighbours graph used as UMAP input}
}
\value{
data.table containing selected 2D embedding, plus annotations per sample
}
\description{
Calculates non-linear embeddings of MMD distances calculated by
calculate_sample_to_sample_MMDs
}
\section{Details}{

\link{Maximum mean discrepancy (MMD) }
\link{Why to centre and not scale}
}

