% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{fit_sampleQC}
\alias{fit_sampleQC}
\title{devtools::load_all('~/work/packages/BayesQC')
devtools::document('~/work/packages/BayesQC')
SampleQC_cells.R
Code to identify outlier cells within sample groups identified by code
in SampleQC_samples. Includes plots of outputs, and plots of fitted
statistics
Fits \code{SampleQC} model to one cluster of samples}
\usage{
fit_sampleQC(
  mmd_list,
  qc_dt,
  qc_names,
  K_list,
  n_cores,
  alpha = 0.01,
  em_iters = 50,
  mcd_alpha = 0.5,
  mcd_iters = 50,
  method = "robust"
)
}
\arguments{
\item{mmd_list}{Outputs from calculate_sample_to_sample_MMDs}

\item{qc_dt}{Data.table of QC metrics for all cells and samples}

\item{qc_names}{List of metrics to actually use for calculating sample-to-sample distances}

\item{K_list}{How many QC celltypes are there for each cluster of samples?}

\item{alpha}{Chi squared threshold to define outliers}

\item{em_iters}{Maximum number of EM iterations}

\item{mcd_alpha, mcd_iters}{Parameters for robust estimation of celltype means and covariances}

\item{method}{Which of various implemented options should be used?}
}
\value{
list, containing MMD values and sample clusters based on MMD values
}
\description{
devtools::load_all('~/work/packages/BayesQC')
devtools::document('~/work/packages/BayesQC')
SampleQC_cells.R
Code to identify outlier cells within sample groups identified by code
in SampleQC_samples. Includes plots of outputs, and plots of fitted
statistics
Fits \code{SampleQC} model to one cluster of samples
}
\section{Details}{

\link{How to define K}
\link{What difference do mcd parameters make?}
}

\keyword{internal}
