% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_sampleqc_mle_cpp}
\alias{fit_sampleqc_mle_cpp}
\title{Fits a multivariate Gaussian mixture model to given data via MLE}
\usage{
fit_sampleqc_mle_cpp(x, init_gamma_i, groups, D, J, K, N, n_iter, seed)
}
\description{
Fits a multivariate Gaussian mixture model to given data via MLE
}
\keyword{internal}
