% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_sampleqc_robust_cpp}
\alias{fit_sampleqc_robust_cpp}
\title{Robustly fits a multivariate Gaussian mixture model to given data}
\usage{
fit_sampleqc_robust_cpp(
  x,
  init_z,
  groups,
  D,
  J,
  K,
  N,
  em_iters,
  mcd_alpha,
  mcd_iters,
  seed,
  track = FALSE
)
}
\description{
Robustly fits a multivariate Gaussian mixture model to given data
}
\keyword{internal}
