% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{get_n_groups}
\alias{get_n_groups}
\title{Extracts number of groups}
\usage{
get_n_groups(qc_obj)
}
\arguments{
\item{qc_obj}{Output from \code{calc_pairwise_mmds}}
}
\value{
number of groups identified by \code{calc_pairwise_mmds}
}
\description{
Extracts number of groups
}
