% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_setters_getters.R
\name{get_outliers}
\alias{get_outliers}
\title{Extracts list of outliers}
\usage{
get_outliers(qc_obj, exc_groups = NULL)
}
\arguments{
\item{qc_obj}{Output from fit_sampleQC}

\item{exc_groups}{List of sample groups to exclude}
}
\value{
\code{data.table} containing cell_id, sample_id and whether outlier
or not
}
\description{
Extracts list of outliers
}
