% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{list_known_metrics}
\alias{list_known_metrics}
\title{Shows the list of QC metrics that for which \pkg{SampleQC} currently has
specific functionality. \pkg{SampleQC} will happily use metrics that aren't
in this list, however for those in this list it can plot a couple of extra
things.}
\usage{
list_known_metrics()
}
\value{
character vector of QC metrics that SampleQC knows about
}
\description{
Shows the list of QC metrics that for which \pkg{SampleQC} currently has
specific functionality. \pkg{SampleQC} will happily use metrics that aren't
in this list, however for those in this list it can plot a couple of extra
things.
}
