% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{make_qc_dt.data.frame}
\alias{make_qc_dt.data.frame}
\title{Checks that input is ok, puts it into expected format}
\usage{
\method{make_qc_dt}{data.frame}(x, qc_names = c("log_counts", "log_feats", "logit_mito"))
}
\arguments{
\item{x}{data.frame object containing calculated QC metrics}

\item{qc_names}{list of qc_names that need to be extracted}
}
\value{
qc_dt, a data.table containing the sample variable plus qc metrics
}
\description{
Checks that input is ok, puts it into expected format
}
