% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_dt.R
\name{make_qc_dt.SingleCellExperiment}
\alias{make_qc_dt.SingleCellExperiment}
\title{Checks that input is ok, puts it into expected format}
\usage{
\method{make_qc_dt}{SingleCellExperiment}(x, qc_names = c("log_counts", "log_feats", "logit_mito"))
}
\arguments{
\item{x}{SingleCellExperiment or data.frame (or data.table, some other class
inheriting data.frame) containing calculated QC metrics}

\item{qc_names}{list of qc_names that need to be extracted}
}
\value{
qc_dt, a data.table containing the sample variable plus qc metrics
}
\description{
Checks that input is ok, puts it into expected format
}
