% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmd_fn}
\alias{mmd_fn}
\title{Calculate MMD between two matrices
“A Kernel Two-Sample Test.” Journal of Machine Learning Research: JMLR 13 (Mar): 723–73.
Gretton, Arthur, Karsten M. Borgwardt, Malte J. Rasch, Bernhard Schölkopf, and Alexander Smola. 2012.}
\usage{
mmd_fn(X, Y, sigma)
}
\arguments{
\item{X}{A matrix of observations (rows = observations)}

\item{Y}{A matrix of observations (rows = observations)}

\item{sigma}{Scale factor for Gaussian kernel)}
}
\description{
Calculate MMD between two matrices
“A Kernel Two-Sample Test.” Journal of Machine Learning Research: JMLR 13 (Mar): 723–73.
Gretton, Arthur, Karsten M. Borgwardt, Malte J. Rasch, Bernhard Schölkopf, and Alexander Smola. 2012.
}
