% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_js_likelihoods.R
\name{plot_alpha_js_likelihoods}
\alias{plot_alpha_js_likelihoods}
\title{plot_alpha_js_likelihoods}
\usage{
plot_alpha_js_likelihoods(qc_obj, sel_group, df = 5)
}
\arguments{
\item{qc_obj}{Output from function fit_sampleqc}

\item{sel_group}{Which sample group to plot}

\item{df}{degrees of freedom for multivariate t}
}
\value{
ggplot object
}
\description{
Plots inferred sample-level statistics, with empirical log-likelihood
}
