% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{plot_beta_ks}
\alias{plot_beta_ks}
\title{Plots estimates of cluster splits by sample, with KL divergence from mean}
\usage{
plot_beta_ks(fit_obj)
}
\arguments{
\item{fit_obj}{fit_obj object from function fit_qc_model}
}
\value{
ggplot object
}
\description{
Plots estimates of cluster splits by sample, with KL divergence from mean
}
