% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Plots non-linear embeddings of MMD distances calculated by
embed_sample_to_sample_MMDs}
\usage{
plot_embeddings(
  qc_obj,
  var_type = c("discrete", "continuous"),
  sel_embed = c("MDS", "UMAP")
)
}
\arguments{
\item{qc_obj}{SampleQC object}

\item{var_type}{Is this annotation discrete or continuous?}

\item{sel_embed}{Which type of embedding to use}
}
\description{
Plots non-linear embeddings of MMD distances calculated by
embed_sample_to_sample_MMDs
}
