% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_over_biaxials.R
\name{plot_fit_over_biaxials}
\alias{plot_fit_over_biaxials}
\title{plot_fit_over_biaxials}
\usage{
plot_fit_over_biaxials(qc_obj, sel_sample, qc_names = NULL, alpha_cut = 0.01)
}
\arguments{
\item{qc_obj}{Output from function fit_sampleqc}

\item{sel_sample}{Which sample to plot?}

\item{qc_names}{List of metrics to be plotted (first is used for y axis).
Default is to use the qc_names specified in metadata(qc_obj)$qc_names; this
option is included to allow for "zooming in" on a subset, or changing the
order in which they are plotted.}

\item{alpha_cut}{Chi-squared quantile to determine how large the ellipses
should be.}
}
\value{
ggplot object
}
\description{
Plots inferred sample-level statistics over QC biaxials
}
