% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mmd_heatmap.R
\name{plot_mmd_heatmap}
\alias{plot_mmd_heatmap}
\title{plot_mmd_heatmap}
\usage{
plot_mmd_heatmap(qc_obj)
}
\arguments{
\item{qc_obj}{Outputs from function calc_pairwise_mmds}
}
\value{
ggplot object
}
\description{
Plots heatmap of sample-sample distances (as measured by MMD)
}
