% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers.R
\name{plot_outliers}
\alias{plot_outliers}
\title{plot_outliers}
\usage{
plot_outliers(qc_obj, sel_sample, outliers_dt = NULL)
}
\arguments{
\item{qc_obj}{Output from function fit_sampleqc}

\item{sel_sample}{Selected sample}

\item{outliers_dt}{Optional alternative outlier specification; see function
\code{get_outliers}.}
}
\value{
ggplot object
}
\description{
Calculates path for an ellipse based on 2D mu and sigma values
}
