% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{plot_outliers_one_sample}
\alias{plot_outliers_one_sample}
\title{Calculates path for an ellipse based on 2D mu and sigma values}
\usage{
plot_outliers_one_sample(em_list, s)
}
\arguments{
\item{em_list}{Output from .fit_one_sampleQC}

\item{s}{Selected sample}
}
\value{
ggplot object
}
\description{
Calculates path for an ellipse based on 2D mu and sigma values
}
