% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC-package.R
\docType{package}
\name{SampleQC}
\alias{SampleQC}
\alias{SampleQC-package}
\title{SampleQC: robust multivariate, multi-celltype, multi-sample quality control for single cell RNA-seq}
\description{
\link{add description}
}
\section{SampleQC samples}{

[]
}

\section{SampleQC outliers}{

[]
}

