% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{sim_experiment}
\alias{sim_experiment}
\title{devtools::load_all('~/work/packages/BayesQC')
devtools::document('~/work/packages/BayesQC')
SampleQC_sims.R
Complex simulations of good and bad quality cells
Simulates QC metrics for whole experiment}
\usage{
sim_experiment(
  n_groups = 4,
  n_cells = 1e+05,
  cells_p_s = 2000,
  D = 3,
  K = 4,
  qc_names = c("log_counts", "log_feats", "logit_mito")
)
}
\arguments{
\item{n_groups}{How many sample groups?}

\item{n_cells}{How many cells in total to simulate}

\item{cells_p_s}{Cells per sample, average value (i.e. default is that each
sample has 2000 cells on average)}

\item{D}{How many QC metrics do you want}

\item{K}{How many mixture components should there be in total? (see Details)}
}
\value{
list with two entries:
\itemize{
\item \code{qc_dt}, a \code{\link[data.table]{data.table}}
\item \code{params}, a list specifying the true parameter values for each group
}
}
\description{
devtools::load_all('~/work/packages/BayesQC')
devtools::document('~/work/packages/BayesQC')
SampleQC_sims.R
Complex simulations of good and bad quality cells
Simulates QC metrics for whole experiment
}
\section{Details}{

\link{describe what the components are}
}

