% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subsample_mmd_fn}
\alias{subsample_mmd_fn}
\title{Calculate mean subsampled MMD between two matrices}
\usage{
subsample_mmd_fn(X, Y, subsample, n_times, sigma)
}
\arguments{
\item{X}{A matrix of observations (rows = observations)}

\item{Y}{A matrix of observations (rows = observations)}

\item{subsample}{How many to subsample from each matrix?}

\item{n_times}{Mean over how many subsamples?}

\item{sigma}{Scale factor for Gaussian kernel}
}
\description{
Calculate mean subsampled MMD between two matrices
}
