% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{calculate_sample_to_sample_MMDs}
\alias{calculate_sample_to_sample_MMDs}
\title{Calculates MMD distances between all samples in qc_dt object, then and
embeds them with some annotations.}
\usage{
calculate_sample_to_sample_MMDs(
  qc_dt,
  qc_names = c("log_counts", "log_feats", "logit_mito"),
  annots_disc = NULL,
  annots_cont = NULL,
  n_cores = 4,
  sigma,
  subsample = 200,
  n_times = 10,
  centre_samples = TRUE,
  scale_samples = FALSE
)
}
\arguments{
\item{qc_dt}{Data.table of QC metrics for all cells and samples}

\item{qc_names}{List of metrics to actually use for calculating
sample-to-sample distances}

\item{n_cores}{How many cores to parallelize over}

\item{sigma}{Scale for MMD kernel (defaults to length of qc_names)}

\item{subsample}{Should we downsample the number of cells per sample to
this number}

\item{n_times}{How many times do we sample MMD between each pair of samples?
(if subsampled, MMD is a stochastic value)}

\item{centre_samples, scale_samples}{Should we centre or scale the values
within each sample before calculating MMD?}
}
\value{
slightly funky \code{sce} object
}
\description{
Calculates MMD distances between all samples in qc_dt object, then and
embeds them with some annotations.
}
\section{Details}{

\link{Maximum mean discrepancy (MMD) }
\link{Why to centre and not scale}
}

