% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.calc_ellipses_dt}
\alias{.calc_ellipses_dt}
\title{Calculates path for an ellipse based on 2D mu and sigma values}
\usage{
.calc_ellipses_dt(fit_obj, sel_sample, dim1, dim2, dim2_name,
  alpha = 0.01)
}
\arguments{
\item{fit_obj}{fit_obj object from function fit_qc_model}

\item{dim1}{dimension to use}

\item{dim2}{dimension to use}

\item{alpha}{contour threshold}
}
\value{
data.table object
}
\description{
Calculates path for an ellipse based on 2D mu and sigma values
}
\keyword{internal}
