% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_over_biaxials.R
\name{.calc_one_ellipse}
\alias{.calc_one_ellipse}
\title{Calculates path for an ellipse based on 2D mu and sigma values}
\usage{
.calc_one_ellipse(mu, sigma, alpha)
}
\arguments{
\item{mu}{mean vector}

\item{sigma}{covariance matrix}

\item{alpha}{contour threshold}
}
\value{
data.table object
}
\description{
Calculates path for an ellipse based on 2D mu and sigma values
}
\keyword{internal}
