% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{.calc_p_jk_proportions}
\alias{.calc_p_jk_proportions}
\title{DEPRECATED: Calculate proportions of cells within each sample allocated
to each}
\usage{
.calc_p_jk_proportions(fit_obj, x, groups, alpha)
}
\arguments{
\item{fit_obj}{Most of output from .fit_one_sampleQC}

\item{x}{Matrix of QC values}

\item{groups}{Sample labels for every cell}

\item{alpha}{Chi squared threshold defining outliers}
}
\value{
data.table containing fitted parameters
}
\description{
DEPRECATED: Calculate proportions of cells within each sample allocated
to each
}
\keyword{internal}
