% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{.draw_dir_0}
\alias{.draw_dir_0}
\title{Determines Dirichlet parameter for p_jk draws}
\usage{
.draw_dir_0(K)
}
\arguments{
\item{K}{number of components}
}
\value{
vector of K columns
}
\description{
Determines Dirichlet parameter for p_jk draws
}
\keyword{internal}
