% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{.process_annot_vars}
\alias{.process_annot_vars}
\title{Checks annotation variables}
\usage{
.process_annot_vars(qc_dt, specified, auto)
}
\arguments{
\item{qc_dt}{\code{data.table} of QC metrics}

\item{specified}{Requested annotation variables}

\item{auto}{Automatically generated annotation variables}
}
\value{
character vector
}
\description{
Checks annotation variables
}
\keyword{internal}
