% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qcs.R
\name{.sim_sample_group}
\alias{.sim_sample_group}
\title{Simulates QC metrics for one group of samples}
\usage{
.sim_sample_group(
  N,
  D,
  J,
  K,
  mu_0,
  beta_k,
  Sigma_k,
  p_out_0,
  theta_0,
  p_loss_0
)
}
\arguments{
\item{D}{number of dimensions}

\item{J}{number of groups}

\item{K}{number of components}
}
\value{
list with all parameters for this group
}
\description{
Simulates QC metrics for one group of samples
}
\section{Details}{

}

\keyword{internal}
