% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{fit_sampleQC}
\alias{fit_sampleQC}
\title{devtools::load_all('~/work/packages/BayesQC')
devtools::document('~/work/packages/BayesQC')
SampleQC_cells.R
Code to identify outlier cells within sample groups identified by code
in SampleQC_samples. Includes plots of outputs, and plots of fitted
statistics
Fits \code{SampleQC} model to one cluster of samples}
\usage{
fit_sampleQC(
  qc_obj,
  K_all = NULL,
  K_list = NULL,
  n_cores,
  alpha = 0.01,
  em_iters = 50,
  mcd_alpha = 0.5,
  mcd_iters = 50,
  method = c("robust", "mle"),
  track = FALSE
)
}
\arguments{
\item{qc_obj}{Output from calculate_sample_to_sample_MMDs}

\item{K_all, K_list}{How many QC celltypes do we expect? Exactly one of K_all
and K_list should be specified. If the user wants to fit the same model to
all samples they should specify \code{K_all} as an integer. If the user wants to
fit a different model to each of the sample groups identified by
\code{calculate_sample_to_sample_MMDs}, they should specify \code{K_list} as an
integer vector with the same number of entries as sample groups. See Details.}

\item{n_cores}{How many cores to use; n_cores=1 runs in serial. The default
depends on whether K_all or K_list is used: if K_all is set, n_cores=1; if
K_list is set, n_cores=length(K_list).}

\item{alpha}{Chi squared threshold to define outliers}

\item{em_iters}{Maximum number of EM iterations}

\item{mcd_alpha, mcd_iters}{Parameters for robust estimation of celltype
means and covariances}

\item{method}{Which of various implemented options should be used?}

\item{track}{Track values of parameters during fitting (used for debugging)}
}
\value{
list, containing MMD values and sample clusters based on MMD values
}
\description{
devtools::load_all('~/work/packages/BayesQC')
devtools::document('~/work/packages/BayesQC')
SampleQC_cells.R
Code to identify outlier cells within sample groups identified by code
in SampleQC_samples. Includes plots of outputs, and plots of fitted
statistics
Fits \code{SampleQC} model to one cluster of samples
}
\section{Details}{


\link{How to define K}

\link[=Start with `K_all=1`]{Start with \code{K_all=1}}

\link{Add more details of K_all vs K_list?}

\link{What difference do mcd parameters make?}
}

