% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{fit_sampleQC}
\alias{fit_sampleQC}
\title{Fits \code{SampleQC} model to one cluster of samples}
\usage{
fit_sampleQC(
  qc_obj,
  K_all = NULL,
  K_list = NULL,
  n_cores = NULL,
  alpha = 0.01,
  em_iters = 50,
  mcd_alpha = 0.5,
  mcd_iters = 50,
  method = c("robust", "mle"),
  bp_seed = 22,
  track = FALSE
)
}
\arguments{
\item{qc_obj}{Output from calculate_sample_to_sample_MMDs}

\item{K_all, K_list}{How many QC celltypes do we expect? Exactly one of K_all
and K_list should be specified. If the user wants to fit the same model to
all samples they should specify \code{K_all} as an integer. If the user wants to
fit a different model to each of the sample groups identified by
\code{calculate_sample_to_sample_MMDs}, they should specify \code{K_list} as an
integer vector with the same number of entries as sample groups. See Details.}

\item{n_cores}{How many cores to use; n_cores=1 runs in serial. The default
depends on whether K_all or K_list is used: if K_all is set, n_cores=1; if
K_list is set, n_cores=length(K_list).}

\item{alpha}{Chi squared threshold to define outliers}

\item{em_iters}{Maximum number of EM iterations}

\item{mcd_alpha, mcd_iters}{Parameters for robust estimation of celltype
means and covariances. \emph{mcd_alpha} is the proportion of the data which
the robust covariance estimator attempts to fit; adjust this to be lower
when more outliers are present. \emph{mcd_iters} is the maximum number of
iterations allowed for estimating the covariance matrix, and should only
rarely be hit.}

\item{method}{Should the fitting be robust, or maximum likelihood
estimation? In general, robust should be used.}

\item{bp_seed}{random seed for BiocParallel workers}

\item{track}{Track values of parameters during fitting (used for debugging)}
}
\value{
list, containing MMD values and sample clusters based on MMD values
}
\description{
Running \code{\link{calculate_sample_to_sample_MMDs()}} identifies groups
of samples with similar distributions of QC metrics ('sample groups').
\code{\link{SampleQC}} then fits a multivariate Gaussian mixture model to
each sample group, and uses these distributions to determine outlier cells
without celltype bias.
}
\details{
Determining the number of components to use in a Gaussian mixture model is
a hard problem that \code{\link{SampleQC}} does not attempt to solve.
However, \code{\link{SampleQC}} provides diagnostic plots that assist the
user to make this decision themself.

We recommend the following workflow. The first step is always to identify
sample groups (and derive embeddings) by calling
\code{\link{calculate_sample_to_sample_MMDs}}. Next, we recommend calling
\code{\link{fit_sampleQC}} with \code{K_list=rep(1, get_n_groups(qc_obj)))},
then rendering outputs with \code{make_SampleQC_report(qc_obj, save_dir, 
'test')}. This fits the simplest possible model to each sample group (i.e.
one component), and renders a report with biaxial plots.

The user can then use the biaxial plots to determine how many components
are appropriate for each sample group. So if
}
