% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{make_qc_dt}
\alias{make_qc_dt}
\title{Makes data.table of specified QC metrics}
\usage{
make_qc_dt(x, qc_names = c("log_counts", "log_feats", "logit_mito"))
}
\arguments{
\item{x}{either SCE, or data.frame object containing calculated QC metrics}

\item{qc_names}{list of qc_names that need to be extracted}
}
\value{
qc_dt, a data.table containing the sample variable plus qc metrics
}
\description{
Extracts a \code{data.table} of QC metrics, either by calculating them from
a SingleCellExperiment object, or copying them from a \code{data.frame} that
the user has already made. Also does a little bit of checking of the inputs.
}
\details{
WARNING: This bit of code definitely needs more testing / thought /
improvements... Please let me know if it falls over!
}
