% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sampleqc_report.R
\name{make_sampleqc_report}
\alias{make_sampleqc_report}
\title{make_sampleqc_report}
\usage{
make_sampleqc_report(qc_obj, save_dir, proj_name)
}
\arguments{
\item{qc_obj}{Outputs from fit_sampleqc}

\item{save_dir}{Directory to save into (must exist)}

\item{proj_name}{Name to use to in file}
}
\description{
Renders SampleQC report into specified format.
}
\section{Details}{
NA
}

