% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_utils.R
\name{make_SampleQC_report}
\alias{make_SampleQC_report}
\title{Renders SampleQC report into specified format.}
\usage{
make_SampleQC_report(qc_obj, save_dir, proj_name)
}
\arguments{
\item{qc_obj}{Outputs from fit_sampleQC}

\item{save_dir}{Directory to save into (must exist)}

\item{proj_name}{Name to use to in file}
}
\description{
Renders SampleQC report into specified format.
}
\section{Details}{
NA
}

