% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{plot_alpha_js}
\alias{plot_alpha_js}
\title{Plots inferred sample-level statistics, with empirical log-likelihood}
\usage{
plot_alpha_js(qc_obj, sel_group, df = 5, qc_idx = c(1, 2), pc_idx = c(1, 2))
}
\arguments{
\item{qc_obj}{Output from function fit_sampleQC}

\item{sel_group}{Which sample group to plot}

\item{df}{Degrees of freedom for multivariate t}

\item{qc_idx}{Which QC metrics to plot?}

\item{pc_idx}{Which PCA components to plot?}
}
\value{
ggplot object
}
\description{
Plots inferred sample-level statistics, with empirical log-likelihood
}
