% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{plot_alpha_js_likelihoods}
\alias{plot_alpha_js_likelihoods}
\title{Plots inferred sample-level statistics, with empirical log-likelihood}
\usage{
plot_alpha_js_likelihoods(fit_obj, df = 5)
}
\arguments{
\item{fit_obj}{fit_obj object from function fit_qc_model}

\item{df}{degrees of freedom for multivariate t}
}
\value{
ggplot object
}
\description{
Plots inferred sample-level statistics, with empirical log-likelihood
}
