% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_ks.R
\name{plot_beta_ks}
\alias{plot_beta_ks}
\title{plot_beta_ks}
\usage{
plot_beta_ks(qc_obj, sel_group)
}
\arguments{
\item{qc_obj}{Output from function fit_sampleqc}

\item{sel_group}{Which sample group to plot}
}
\value{
ggplot object
}
\description{
Plots estimates of cluster splits by sample, with KL divergence from mean
}
