% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Plots non-linear embeddings of MMD distances calculated by
embed_sample_to_sample_MMDs}
\usage{
plot_embeddings(mmd_list, var_type, sel_embed)
}
\arguments{
\item{var_type}{Is this annotation discrete or continuous?}

\item{sel_embed}{Which type of embedding to use}

\item{embed_dt}{Output from embed_sample_to_sample_MMDs}

\item{annot_vars}{Which set of annotation variables to plot}
}
\description{
Plots non-linear embeddings of MMD distances calculated by
embed_sample_to_sample_MMDs
}
