% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_cells.R
\name{plot_fit_over_biaxials_one_sample}
\alias{plot_fit_over_biaxials_one_sample}
\title{Plots inferred sample-level statistics over QC biaxials}
\usage{
plot_fit_over_biaxials_one_sample(
  fit_obj,
  qc_dt,
  sel_sample,
  qc_names,
  alpha_cut = 0.01
)
}
\arguments{
\item{fit_obj}{fit_obj object from function fit_qc_model}

\item{qc_dt}{data.table of QC metrics}

\item{sel_sample}{which sample to plot?}

\item{qc_names}{list of metrics to be plotted (first is used for y axis)}

\item{alpha_cut}{contour value for ellipses}

\item{fit_list}{list of fit_obj objects (for comparing multiple models)}
}
\value{
ggplot object
}
\description{
Plots inferred sample-level statistics over QC biaxials
}
