% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{plot_mmd_heatmap}
\alias{plot_mmd_heatmap}
\title{Plots heatmap of sample-sample distances (as measured by MMD)}
\usage{
plot_mmd_heatmap(qc_obj)
}
\arguments{
\item{qc_obj}{Outputs from function calculate_sample_to_sample_MMDs}
}
\value{
ggplot object
}
\description{
Plots heatmap of sample-sample distances (as measured by MMD)
}
