% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_samples.R
\name{plot_mmd_heatmap}
\alias{plot_mmd_heatmap}
\title{Calculates non-linear embeddings of MMD distances calculated by
calculate_sample_to_sample_MMDs}
\usage{
plot_mmd_heatmap(mmd_list)
}
\arguments{
\item{mmd_list}{Outputs from function calculate_sample_to_sample_MMDs}

\item{qc_dt}{Data.table of QC metrics for all cells and samples}

\item{annot_discrete}{List of discrete-valued variables to plot later}

\item{annot_cont}{List of continuous-valued variables to plot later}

\item{n_nhbrs}{How many neighbours in UMAP nearest neighbours graph}
}
\value{
ggplot object
}
\description{
Calculates non-linear embeddings of MMD distances calculated by
calculate_sample_to_sample_MMDs
}
\section{Details}{

\link{Maximum mean discrepancy (MMD) }
\link{Why to centre and not scale}
}

