% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC-package.R
\docType{package}
\name{SampleQC}
\alias{SampleQC}
\title{SampleQC: robust multivariate, multi-celltype, multi-sample quality control for single cell RNA-seq}
\description{
\code{SampleQC} is an R package for robust multivariate, multi-
celltype, multi-sample quality control for single cell RNA-seq. Unimodal
approaches to QC, such as \code{scater}, may preferentially remove healthy cells
from celltypes with extreme QC statistics, introducing bias. \code{SampleQC}
removes this bias by fitting a Gaussian mixture model. The fit is done
across multiple celltypes simultaneously, and using robust estimators
allows it to find the distributions in the (expected) presence of outliers.
}
