% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleQC_sims.R
\name{sim_experiment}
\alias{sim_experiment}
\title{Simulates QC metrics for whole experiment}
\usage{
sim_experiment(
  n_groups = 4,
  n_cells = 1e+05,
  cells_p_s = 2000,
  D = 3,
  K = 4,
  qc_names = c("log_counts", "log_feats", "logit_mito"),
  sel_ks = NULL
)
}
\arguments{
\item{n_groups}{How many sample groups?}

\item{n_cells}{How many cells in total to simulate}

\item{cells_p_s}{Cells per sample, average value (i.e. default is that each
sample has 2000 cells on average)}

\item{D}{How many QC metrics do you want}

\item{K}{How many mixture components should there be in total? (see Details)}

\item{qc_names}{Names for QC metrics}
}
\value{
list with multiple entries:
\itemize{
\item qc_ok: \code{data.table} of cell QC metrics \emph{before} outlier
perturbation
\item qc_out: \code{data.table} of cell QC metrics \emph{after} outlier
perturbation
\item x_ok, x_out: matrices of values in qc_ok, qc_out respectively
\item groups: vector of true sample groups
\item samples: vector of sample_ids
\item z: vector of true celltype / mixture component values
\item outliers: vector of outlier status (0/1)
\item group_sims: detailed list of simulation outputs for each sample group
\item expt_params: list of whole experiment-level parameters (e.g. celltype
means and covariance matrices)
}
}
\description{
Detailed simulation of QC metric distribution across whole experiment,
assuming multiple celltypes, which are shared across different groups of
samples ('sample groups'). Parameters are derived from real data and are
hopefully reasonably realistic.
}
\details{
The simulation first randomly generates \emph{K} celltypes (= mixture
components), each with its own mean and covariance matrix. It also
generates parameters for \emph{n_groups}, including: how many cells in each
group; how many samples; which celltypes are present in that group;
hyperparameters for outliers in that group.

Given these generated quantities, a set of QC metric vectors representing
cells is drawn for each sample group, and combined into \emph{qc_dt}. Each
row in this \code{data.table} represents a cell, and has annotations showing
}
